+++ CHANGELOG.md
@@ -5,12 +5,14 @@ Bob versions changelog
 ## [Unreleased]
 #### Added
 - If no credentials provided and default user specified in config, then request will have default user permissions (#437)
+- User's permissions can be set via role and/or claims (#408)
 
 #### Changed
+- Move files for linux packages into separate directory (#534)
 
 
 #### Fixed
-
+- "CredentialsNotProvided", "UserNotFound", "UnauthorizedRequest" now have "Unauthenticated" code (#528)
 
 #### Updated