+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Bob versions changelog
 
 ## [Unreleased]
 #### Added
+- SHA512 hash of password with salt 'bob' can be specified instead of password (#304)
 - User's permissions can be set via role and/or claims (#408)
 
 #### Changed