+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Bob versions changelog
 - If no credentials provided and default user specified in config, then request will have default user permissions (#437)
 - SHA512 hash of password with salt 'bob' can be specified instead of password (#304)
 - User's permissions can be set via role and/or claims (#408)
+- Check correctness tool, bobt (#542)
 
 #### Changed
 - Move files for linux packages into separate directory (#534)