+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Bob versions changelog
 
 ## [Unreleased]
 #### Added
+- Add occupied disk space to api and metrics (#501)
 - If no credentials provided and default user specified in config, then request will have default user permissions (#437)
 - SHA512 hash of password with salt 'bob' can be specified instead of password (#304)
 - User's permissions can be set via role and/or claims (#408)
@@ -13,9 +14,10 @@ Bob versions changelog
 - Move files for linux packages into separate directory (#534)
 - Defer index offload on deletion (#363)
 
-
 #### Fixed
+- Fixed internode authorization error (#530)
 - "CredentialsNotProvided", "UserNotFound", "UnauthorizedRequest" now have "Unauthenticated" code (#528)
+- Added conversion of "unauthorized" status into internal error (#540)
 
 #### Updated